% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastafolder.R
\name{aptly}
\alias{aptly}
\title{Aptly Characterize Aptamers}
\usage{
aptly(fasta, cores = 1)
}
\arguments{
\item{fasta}{A multi-sequence FASTA file.}

\item{cores}{The number of cores to use.}
}
\value{
A data.frame of characteristics.
}
\description{
This function procedurally characterizes the secondary structure of RNA sequences
 using regular expressions. The characterizations provided may help identify
 potentially functional aptamers.
}
\details{
This function requires a multi-sequence FASTA file as input. It also depends on the UNIX
 programs RNAfold and imagemagick. Install RNAfold through the ViennaRNA package.
}
