% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{asTempObj}
\alias{asTempObj}
\title{Set Temporary Object}
\usage{
asTempObj(value, pos = 1, envir = as.environment(pos))
}
\arguments{
\item{value}{An object to assign as a temporary object.}

\item{pos}{Where to do the assignment (relative to the \code{asTempObj}
environment). Defaults to \code{pos = 1}.}

\item{envir}{The environment to use.}
}
\value{
The name of the temporary object as assigned in the provided
 environment.
}
\description{
Set a given value to a temporary object in a specified environment.
}

