% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache.key}
\alias{cache.key}
\title{Generate Cache Key}
\usage{
cache.key(...)
}
\arguments{
\item{...}{Any number of objects used to make a key.}
}
\value{
A unique key.
}
\description{
Generates a unique hash key from a list of objects.
}
\seealso{
\code{\link{cache.save}}, \code{\link{cache.load}}
}

