% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache.load}
\alias{cache.load}
\title{Load Value from Cache}
\usage{
cache.load(key, where = paste0(get(), "/.cache"))
}
\arguments{
\item{key}{A unique key for the cache.}

\item{where}{The cache file.}
}
\value{
The imported value.
}
\description{
Loads a value in raw from an SQLite database under a specified key.
}
\seealso{
\code{\link{cache.key}}, \code{\link{cache.save}}
}

