% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache.save}
\alias{cache.save}
\title{Save Value to Cache}
\usage{
cache.save(key, value, where = paste0(getwd(), "/.cache"))
}
\arguments{
\item{key}{A unique key for the cache.}

\item{value}{The value to associate with this key.}

\item{where}{The cache file.}
}
\description{
Saves a value in raw to an SQLite database under a specified key.
}
\seealso{
\code{\link{cache.key}}, \code{\link{cache.load}}
}

