% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ctd}
\alias{ctd}
\title{Comparative Toxicogenomics Database}
\format{A \code{data.frame} with >1,000,000 chemical-gene interactions (rows)
 and 11 annotations (columns). Retrieved on April 9, 2019.}
\source{
\url{http://ctdbase.org/reports/CTD_chem_gene_ixns.csv.gz}
}
\usage{
data(ctd)
}
\description{
These data catalog chemical-gene interactions curated from the literature.
 Each interaction in this database contains useful information about the
 source of this knowledge and the type of interaction observed.
}
\examples{
\dontrun{
library(miSciTools)
data(ctd)
head(ctd[, c("GeneID", "ChemicalName")])
genes <- ctd$GeneID[ctd$ChemicalName == "Platinum"]
pvals <-
  simpliGSEA(genes = genes, # Test "Platinum" interactions
             universe = unique(ctd$GeneID),
             annot.genes = ctd$GeneID,
             annot.terms = ctd$ChemicalName)
head(sort(p.adjust(pvals)))
}
}
\keyword{datasets}
