% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ctd.wide}
\alias{ctd.wide}
\title{Comparative Toxicogenomics Database (Wide Format)}
\format{A \code{data.frame} with >1,000,000 chemical-gene interactions (rows)
 and 11 annotations (columns). Retrieved on April 9, 2019.}
\source{
\url{http://ctdbase.org/reports/CTD_chem_gene_ixns.csv.gz}
}
\usage{
data(ctd.wide)
}
\description{
Unlike \code{data(ctd)}, this object explodes the "InteractionActions"
 column into multiple columns, each describing an association type.
}
\keyword{datasets}
