% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miSciTools.R
\name{demand}
\alias{demand}
\title{Load or Install Package}
\usage{
demand(package)
}
\arguments{
\item{package}{A character string. The package to load or install.}
}
\description{
This function finds and installs a package, if not already installed.
}

