% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aptly.R
\name{fastafolder}
\alias{fastafolder}
\title{Sequence Dissimilarity by Secondary Structure}
\usage{
fastafolder(fasta, rmTails = FALSE, estimateConvergence = FALSE,
  gapOpening = 10, gapExtension = 4, groupBy = 0.5, cores = 1)
}
\arguments{
\item{fasta}{A multi-sequence FASTA file.}

\item{rmTails}{Toggles whether to remove trailing "..." from start and end
of the secondary structure bracket notation.}

\item{estimateConvergence}{Toggles whether to estimate secondary convergence
(log of the secondary alignment score divided by primary alignment score).
Produces intermediate files.}

\item{gapOpening, gapExtension}{Penalty for alignment gaps.
Passed to \code{Biostrings::pairwiseAlignment}.}

\item{groupBy}{The height of the tree at which to cut when defining clusters.
If \code{groupBy >= 1}, defines the total number of clusters to extract.
Disable by setting \code{groupBy = 0}.
Produces intermediate files.}

\item{cores}{The number of cores to use.}
}
\value{
A dissimilarity measure from [0, 1].
}
\description{
This function calculates the dissimilarity between sequences based on secondary structure.
 If \code{estimateConvergence = TRUE}, this function also calculates the dissimilarity
 based on the primary sequence and saves a convergence measure as output. If
 \code{groupBy > 0}, this function also clusters sequences by secondary structure and
 saves additional files as output (including secondary structure figures).
}
\details{
This function requires a multi-sequence FASTA file as input. It also depends on the UNIX
 programs RNAfold and imagemagick. Install RNAfold through the ViennaRNA package.
}
