% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig.R
\name{fig.3D}
\alias{fig.3D}
\title{Make 3D Plot}
\usage{
fig.3D(data, x = 1, y = 2, z = 3, n.group = 1, main = "")
}
\arguments{
\item{data}{A \code{data.frame}.}

\item{x, y, z}{A numeric. The x, y, and z dimensions to plot.}

\item{n.group}{A character vector or factor. Groupings for the points.
Used to color the figure.}

\item{main}{A string. The plot title.}
}
\description{
Make 3D Plot
}
\examples{
\dontrun{
fig.quick3D(mtcars, x = 1, y = 2, z = 3, n.group = mtcars$cyl)
}
}
