% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig.R
\name{fig.cbind}
\alias{fig.cbind}
\title{Vertical Join Figures}
\usage{
fig.cbind(plot1, plot2, file = "fig", dir = getwd(), width = 5,
  height = 5, res = 600)
}
\arguments{
\item{plot1, plot2}{The figures.}

\item{file}{A string. The file to save.}

\item{dir}{A string. The directory.}

\item{width}{The figure dimensions.}

\item{height}{The figure dimensions.}

\item{res}{The figure resolution.}
}
\description{
Vertical Join Figures
}
\examples{
\dontrun{
fig.cbind(plot(1:5, 1:5), plot(1:5, 5:1), file = "Figure2")
}
}
