% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig.R
\name{fig.fromTable}
\alias{fig.fromTable}
\title{Make Figure from Table}
\usage{
fig.fromTable(mat, colmat = matrix("black", nrow(mat), ncol(mat)))
}
\arguments{
\item{mat}{The matrix.}

\item{colmat}{A matrix of colors. Specifies the colors
for each cell in the matrix.}
}
\value{
A grob.
}
\description{
Make Figure from Table
}
\examples{
colmat <- matrix("red", 5, 5)
colmat[,1] <- "blue"
fig.fromTable(iris[1:5, 1:5], colmat)
}
