% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig.R
\name{fig.label}
\alias{fig.label}
\title{Label Figure Panel}
\usage{
fig.label(plot, label = "1) Place Label Here")
}
\arguments{
\item{plot}{The figure.}

\item{label}{The label.}
}
\value{
A grob.
}
\description{
This function adds a label to the top-left of the figure.
}
\examples{
grob <- fig.fromTable(iris[1:5,1:5])
fig.label(grob, "1A) An example")
}
