% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig.R
\name{fig.pretty}
\alias{fig.pretty}
\title{Make \code{ggplot} Pretty}
\usage{
fig.pretty(ggplot, col = FALSE, rotate.x = FALSE)
}
\arguments{
\item{ggplot}{A \code{ggplot} object.}

\item{col}{A boolean. Toggles whether to color the figure with viridis.}

\item{rotate.x}{A boolean. Toggles whether to rotate the x-axis.}
}
\description{
Make \code{ggplot} Pretty
}
\examples{
\dontrun{
p <- ggplot(mtcars, aes(wt, mpg, col = cyl))
p <- p + geom_point()
fig.pretty(p, col = TRUE)
}
}
