% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig.R
\name{fig.rbind}
\alias{fig.rbind}
\title{Horizontal Join Figures}
\usage{
fig.rbind(plot1, plot2, file = "fig", width = 5, height = 5,
  res = 600)
}
\arguments{
\item{plot1, plot2}{The figures.}

\item{file}{A string. The file to save.}

\item{width}{The figure dimensions.}

\item{height}{The figure dimensions.}

\item{res}{The figure resolution.}
}
\description{
Horizontal Join Figures
}
\examples{
\dontrun{
fig.rbind(plot(1:5, 1:5), plot(1:5, 5:1), file = "Figure3")
}
}
