% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig.R
\name{fig.save}
\alias{fig.save}
\title{Save Figure}
\usage{
fig.save(plot, file = "fig", width = 5, height = 5, res = 600)
}
\arguments{
\item{plot}{The figure.}

\item{file}{A string. The file to save.}

\item{width, height}{The figure dimensions.}

\item{res}{The figure resolution.}
}
\description{
This function saves a figure.
}
\examples{
\dontrun{
fig.save(plot(1:5, 1:5), file = "Figure1")
}
}
