% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig.R
\name{fig.tile}
\alias{fig.tile}
\title{Tile Join Figures}
\usage{
fig.tile(..., tile = "2x2", file = "fig", dir = getwd(), width = 5,
  height = 5, res = 600)
}
\arguments{
\item{...}{The figures.}

\item{tile}{A string. Describes how to tile the figures (e.g., "1x3").}

\item{file}{A string. The file to save.}

\item{dir}{A string. The directory.}

\item{width}{The figure dimensions.}

\item{height}{The figure dimensions.}

\item{res}{The figure resolution.}
}
\description{
Tile Join Figures
}
\examples{
\dontrun{
fig.tile(plot(1:5), plot(6:10), plot(11:15), tile = "1x3", file = "Figure 4")
}
}
