% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRecount.R
\name{getRecount}
\alias{getRecount}
\title{Get recount2 Data}
\usage{
getRecount(ID)
}
\arguments{
\item{ID}{A character string. The recount2 ID.}
}
\value{
Saves data to working directory.
}
\description{
This function retrieves data from the recount2 server.
}
