% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig.R
\name{\%+\%}
\alias{\%+\%}
\title{Paste with Sum}
\usage{
a \%+\% b
}
\arguments{
\item{a, b}{A string.}
}
\value{
A string.
}
\description{
Use %+% to paste0 two strings.
}
\examples{
"Now you can paste() like " \%+\% " in Python."
}
