% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{hgnc2liger}
\alias{hgnc2liger}
\title{Quickly Run a Gene Set Enrichment}
\usage{
hgnc2liger(hgnc, score, minK = 10)
}
\arguments{
\item{hgnc}{A vector of gene symbols.}

\item{score}{A vector of scores.}

\item{minK}{An integer. The minimum number of nodes
to which a concept maps to be included in final table.
Skip with minK = 0.}
}
\value{
A \code{data.frame} of results.
}
\description{
Quickly Run a Gene Set Enrichment
}
