% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{insert}
\alias{insert}
\title{Insert New Line}
\usage{
insert(file, after, what)
}
\arguments{
\item{file}{A character string. The text file to modify.}

\item{after}{A character string. A fragment of existing text that marks
the line after which to insert the new text.}

\item{what}{A character string. The new text to insert.}
}
\description{
Inserts a new line of text into an existing file at a specified breakpoint.
}

