% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kegg}
\alias{kegg}
\title{KEGGREST Database}
\format{A \code{data.frame} with 27,944 KEGGREST-gene relationships (rows).
 Retrieved on July 28, 2017.}
\source{
\url{http://rest.kegg.jp/link/pathway/hsa}
}
\usage{
data(kegg)
}
\description{
These data catalog KEGGREST-gene relationships from the KEGGREST database.
}
\keyword{datasets}
