% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miSciTools.R
\name{lift}
\alias{lift}
\title{Lift Over Genome Build}
\usage{
lift(object, from = "hg18", to = "hg19", flatGrl = TRUE)
}
\arguments{
\item{object}{A \code{GRanges} object.}

\item{from}{A character string. The initial UCSC equivalent genome build
(e.g., hg18).}

\item{to}{A character string. The target UCSC equivalent genome build
(e.g., hg19).}

\item{flatGrl}{A logical scalar. Toggles whether to collapse the
default \code{GRangesList} output to a \code{GRanges} output.}
}
\value{
A \code{GRangesList} object or \code{GRanges} object.
}
\description{
This function simplifies the process of "lifting over" coordinates from one
 genome build to another by automatically downloading the necessary
 \code{.over.chain.gz} file.
}
\details{
This function will convert any non-UCSC coordinates to UCSC equivalents.
 Make sure the arguments \code{from} and \code{to} refer to valid UCSC
 genome builds. Feature names will get lost during 'lift over'. To retain,
 these labels, consider storing names as a separate column in the
 \code{GRanges} object.
}

