% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{linc2go}
\alias{linc2go}
\title{Linc2GO GeneSetCollection}
\format{A \code{GeneSetCollection} object from the \code{GSEABase} package.}
\source{
\url{http://www.bioinfo.tsinghua.edu.cn/~liuke/Linc2GO/}
}
\usage{
data(linc2go)
}
\description{
These data expose the Linc2GO database as a \code{GeneSetCollection} object
 to use in conjunction with the \code{Category} and \code{GOstats}
 packages for gene set enrichment analysis. The gene ontology terms
 in this database stem from an application of the competing endogenous
 RNA (ceRNA) hypothesis which speculates that lncRNAs regulate expression,
 at least in part, by competing with mRNAs for microRNA binding sites.
 Since microRNAs bind mRNA transcriptS to silence their expression,
 a microRNA-lncRNA complex could free up that mRNA for expression.
 Therefore, the curators integrate microRNA-mRNA and microRNA-lncRNA
 binding patterns to predict potential lncRNA-mRNA collaborations.
 Based on the gene ontology of the mRNA transcript pairs, the curators can
 make an educated guess about the gene ontology of the lncRNA.
}
\details{
This database uses the lincRNA ID archived by the Human lincRNA Catalog.
 Conveniently, these represent the same terms used by UCSC in the
 \code{TxDb.Hsapiens.UCSC.hg19.lincRNAsTranscripts} \code{TxDb} object.
}
\examples{
\dontrun{
library(miSciTools)
data(linc2go)
show(linc2go)
universe <- unique(unlist(geneIds(linc2go))) # Pull out all features
genes <- geneIds(linc2go)[[2]] # Pull out all GO:0000003 terms
params <-
  Category::GSEAGOHyperGParams(name = "ExampleGSEA",
                               geneSetCollection = linc2go,
                               geneIds = genes,
                               universeGeneIds = universe,
                               ontology = "BP",
                               pvalueCutoff = 0.05,
                               conditional = FALSE,
                               testDirection = "over")
Over <- GOstats::hyperGTest(params)
head(summary(Over))
}
}
\keyword{datasets}

