% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{packageCheck}
\alias{packageCheck}
\title{Check for Packages}
\usage{
packageCheck(packages)
}
\arguments{
\item{packages}{A character vector of packages.}
}
\description{
Triggers an error if a package is not installed.
}
