% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miSciTools.R
\name{peakRAM}
\alias{peakRAM}
\title{Calculate Peak RAM Used}
\usage{
peakRAM(fx)
}
\arguments{
\item{fx}{A function to call (e.g., \code{peakRAM(function() phit(counts))}).}
}
\description{
This function calculates the peak amount of RAM used during a function call.
}

