% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miSciTools.R
\name{qsub}
\alias{qsub}
\title{Qsub Linux Command}
\usage{
qsub(cmd, ...)
}
\arguments{
\item{cmd}{A character string. A Linux command to \code{qsub} or the
location of an R script to \code{qsub}.}

\item{...}{Any additional PBS argument(s). Each argument should
get named according to the bash character argument (e.g., set
\code{-M thom@tpq.me} with \code{qsub(command, M = "thom@tpq.me")}).}
}
\description{
This function sends a Linux command to the PBS queue via \code{qsub}.
}

