% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dex.R
\name{quick.anova}
\alias{quick.anova}
\title{Run Quick ANOVA}
\usage{
quick.anova(x, y)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A factor vector.}
}
\value{
A p-value.
}
\description{
This function runs an ANOVA using f(x, y) syntax instead of
 f(x~y) syntax.
}
