% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dex.R
\name{quick.DESeq2}
\alias{quick.DESeq2}
\title{Run DESeq2 LRT or Wald Test}
\usage{
quick.DESeq2(x, y, formula, test = "Wald")
}
\arguments{
\item{x}{A data.frame of raw sequence counts.}

\item{y}{A data.frame of annotations.}

\item{formula}{The formula used for differential expression analysis.}

\item{test}{The test to run. Choose between "LRT" and "Wald".}
}
\description{
Run DESeq2 LRT or Wald Test for differential expression analysis.
}
