% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand.R
\name{randBoots}
\alias{randBoots}
\title{Randomly Generate Bootstraps}
\usage{
randBoots(n.samples, b = 10, percent.include = 67)
}
\arguments{
\item{n.samples}{The number of samples to group.}

\item{b}{The number of folds to use.}

\item{percent.include}{The number of samples to include.}
}
\description{
This function samples each sample index (from 1 to \code{n.samples})
 into a group (from 1 to \code{b}).
}
