% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand.R
\name{randFolds}
\alias{randFolds}
\title{Randomly Generate Folds}
\usage{
randFolds(n.samples, f = 10)
}
\arguments{
\item{n.samples}{The number of samples to group.}

\item{f}{The number of folds to use.}
}
\description{
This function assigns each sample index (from 1 to \code{n.samples})
 into a fold (from 1 to \code{f}).
}
