% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand.R
\name{randSamps}
\alias{randSamps}
\title{Randomly Assign Class Labels}
\usage{
randSamps(n.samples, n.groups = 2, prob = NULL)
}
\arguments{
\item{n.samples}{The number of samples to label.}

\item{n.groups}{The number unique classes.}

\item{prob}{Argument passed to \code{sample}.}
}
\description{
This function randomly assigns class labels.
}
