% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rda.R
\name{rda.reconstruct}
\alias{rda.reconstruct}
\title{Reconstruct Data from RDA Object}
\usage{
rda.reconstruct(rda)
}
\arguments{
\item{rda}{An object from \code{vegan::rda}.}
}
\value{
Returns a \code{data.frame} object.
}
\description{
This function reconstructs the original data from the
 \code{$CA} slot of a \code{vegan::rda} object. This should
 work for partial RDA also, returning the data "X" that
 excludes the constrained part "Z".
}
