% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replaceNA.R
\name{replaceNA}
\alias{replaceNA}
\title{Replace NA Values}
\usage{
replaceNA(x)
}
\arguments{
\item{x}{A vector or \code{data.frame}.}
}
\description{
This function replaces all NAs in a vector or \code{data.frame}.
 If the vector (or column) is numeric, NAs get replaced by the median.
 If the vector (or column) is a character or factor, NAs get
 replaced by the mode.
}
