% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miSciTools.R
\name{simpliGSEA}
\alias{simpliGSEA}
\title{Perform Gene Set Enrichment Analysis}
\usage{
simpliGSEA(genes, universe, annot.genes, annot.terms)
}
\arguments{
\item{genes}{A character vector. The gene IDs of interest.}

\item{universe}{A character vector. The complete set of gene IDs used to
generate the gene IDs of interest.}

\item{annot.genes}{A character vector. The gene IDs to which \code{annot.terms}
will correspond.}

\item{annot.terms}{A character vector. The annotations corresponding to
\code{annot.genes}.}
}
\value{
A vector of p-values named based on the tested annotation.
}
\description{
This function simplifies gene set enrichment analysis for custom annotations.
 Do not use this function for annotations that have directed acyclic graph
 (DAG) relationships (e.g., GO). Gene set enrichment analysis performed
 using Fisher's exact test for count data. This function will implicitly
 filter non-unique \code{annot.genes} and \code{annot.terms} pairs.
}

