% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test.lequal}
\alias{test.lequal}
\title{Test All Equal Within List}
\usage{
test.lequal(list)
}
\arguments{
\item{list}{A list.}
}
\description{
This function tests whether all elements in a list are identical.
 Works like an iterative \code{all.equal}.
}
