% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miSciTools.R
\name{writeR}
\alias{writeR}
\title{Write R Script}
\usage{
writeR(R, folder = tempdir(), file = paste0(basename(folder), ".R"))
}
\arguments{
\item{R}{A character string. The R script to save.}

\item{folder}{A character string. The folder in which to save the script.
Defaults to a temporary directory name.}

\item{file}{A character string. The file name to assign to the script.
Defaults to a temporary file name.}
}
\value{
The file path for the saved R script.
}
\description{
This function saves a character string as an R script. By default, this
 script gets stored in a temporary directory as a temporarily file.
 This function returns the file path for the saved R script for []
 use (e.g., via \code{\link{qsub}}).
}

