% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tt.R
\name{wwtest}
\alias{wwtest}
\title{Make a Table of Wilcoxon Rank Sum test Results}
\usage{
wwtest(df, y, x, subset, digits = 2, ...)
}
\arguments{
\item{df}{A tidy long-form \code{data.frame}.}

\item{y}{A string. The column name of the outcome variable.}

\item{x}{A string. The column name of the contrast variable.}

\item{subset}{A boolean of length \code{nrow(df)}. How to
subset the \code{data.frame} prior to \code{t.test}.}

\item{digits}{An integer. The number of significant digits.}

\item{...}{Arguments passed to \code{t.test}.}
}
\value{
A table of Wilcoxon Rank Sum test results.
}
\description{
Make a Table of Wilcoxon Rank Sum test Results
}
