% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastafolder.R
\name{aptly}
\alias{aptly}
\title{Aptly Characterize Aptamers}
\usage{
aptly(fasta)
}
\arguments{
\item{fasta}{A multi-sequence FASTA file.}
}
\value{
A data.frame of characteristics.
}
\description{
This function procedurally characterizes the secondary structure of RNA sequences
 using regular expressions. The characterizations used may have relevance to identifying
 potentially functional aptamers.
}
\details{
This function requires a multi-sequence FASTA file as input. It also depends on the UNIX
 programs RNAfold and imagemagick. Install RNAfold through the ViennaRNA package.
}
