% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aptly.R
\name{aptly}
\alias{aptly}
\title{Aptly Characterize Aptamers}
\usage{
aptly(fasta, RNAfold = "RNAfold", cores = 1, select)
}
\arguments{
\item{fasta}{A multi-sequence FASTA file.}

\item{RNAfold}{The commandline program to call. Append RNAfold arguments here.}

\item{cores}{The number of cores to use.}

\item{select}{A numeric or character vector. Selects which sequences from
the multi-sequence FASTA file to include in the analysis.}
}
\value{
A list of characteristics.
}
\description{
This function procedurally characterizes the secondary structure of RNA sequences
 using regular expressions. The characterizations provided may help identify
 potentially functional aptamers.
}
\details{
This function requires a multi-sequence FASTA file as input. It also depends on the UNIX
 programs RNAfold and imagemagick. Install RNAfold through the ViennaRNA package.
}
