% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{demand}
\alias{demand}
\title{Load or Install Package}
\usage{
demand(packages)
}
\arguments{
\item{packages}{A character vector. The package(s) to load or install.}
}
\description{
This function installs a package if not already installed, then loads it.
}
