% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dex.R
\name{quick.edgeR}
\alias{quick.edgeR}
\title{Run edgeR Exact Test}
\usage{
quick.edgeR(counts, group)
}
\arguments{
\item{counts}{A data.frame of raw sequence counts.}

\item{group}{A character vector of group labels.}
}
\description{
Run edgeR Exact Test for differential expression analysis.
}
