% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miSciTools.R
\name{writeR}
\alias{writeR}
\title{Write R Script}
\usage{
writeR(..., file = paste0(tempfile(), ".R"), preview = FALSE)
}
\arguments{
\item{...}{Any number of character strings or R expressions to join
together and save as an R script.}

\item{file}{A character string. The file path, including directory, where
to save the new script. Defaults to a temporary file name.}

\item{preview}{A logical scalar. Toggles whether to preview the script
in the console before saving it. Defaults to \code{FALSE}.}
}
\value{
The file path for the saved R script.
}
\description{
This function saves a character string as an R script. By default, this
 script gets stored in a temporary directory as a temporarily file.
 This function returns the file path for the saved R script for later
 use (e.g., via \code{\link{qsub}}).
}

